ai_email_automation/requirements.txt
openai>=0.27.0
APScheduler>=3.9.1
requests>=2.25.1
python-dotenv>=0.21.0